+++ src/models/employeeAssistanceModels.mjs
@@ -1,0 +1,22 @@
+import { DataTypes } from 'sequelize'
+import { sequelize } from '../../config/db_config.mjs'
+
+sequelize.define('employee_assistance', {
+  id: {
+    type: DataTypes.INTEGER,
+    primaryKey: true,
+    autoIncrement: true
+  },
+  start_date: {
+    type: DataTypes.DATE,
+    allowNull: false
+  },
+  end_date: {
+    type: DataTypes.DATE,
+    allowNull: false
+  },
+  absences: {
+    type: DataTypes.INTEGER,
+    allowNull: false
+  }
+})
